Add Css/Js to profile page
Create profile.scss which manage
the CSS for profile page.
Create profile.js which manage
the JS for profile page.
Modify ProfileUserType.php to be more
precise, usefull and
do the task it as to do.
Modify profile.html.twig to add
CSS to it.
Modify webpack.config.js, so that
profile.scss and profile.js
are usable in the twig file.